reactive_service/requirements.txt
sanic
kafka-python
gunicorn